refactor(thresh): add separate threshold mod
